8280437: Move G1BufferNodeList to gc/shared
